requirements.txt
apache-airflow==2.5.2
apache-airflow-providers-amazon
pyspark
minio
bs4
scikit-learn==0.24.2
matplotlib==3.3.3
